tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/url/002.html

<!doctype html>
<title>WebSockets: setting url</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
test(function() {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
  ws.url = SCHEME_DOMAIN_PORT+'/test';
  assert_equals(ws.url, SCHEME_DOMAIN_PORT+'/');
});
</script>
